* Added x64 config * Split profi_decls between decls for the client and build options for the library * Added Pause/Resume profiling API

+++ include/IAllocator.h
@@ -44,7 +44,7 @@ private:
 	CountMallocPolicy(const CountMallocPolicy&);
 	CountMallocPolicy& operator=(const CountMallocPolicy&);
 
-	unsigned m_Allocated;
+	size_t m_Allocated;
 };
 
 class NoCountPolicy
